requirements.txt
argcomplete==1.11.1
asgiref==3.2.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.2
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.7.0
jsonpath-rw==1.4.0
jsonschema==2.6.0
more-itertools==5.0.0
mysqlclient==1.4.6
ply==3.11
prettytable==0.7.2
prompt-toolkit==1.0.15
psycopg2==2.8.5
pycparser==2.20
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
sseclient-py==1.7
urllib3==1.25.8
wcwidth==0.2.5
whitenoise==5.0.1
zipp==1.0.0
